aaaaaa

+++ DTO/StocksDTO.cs
@@ -3,7 +3,7 @@ namespace DTO;
 public class StocksDTO
 {
 
-    public Double quantity;
+    public int quantity;
     public Double unit_price;
 
     public ProductDTO ProductDTO;
